Load and manipulate proff data in db

+++ server/db/dbfuncs/regexTest.js
@@ -9,4 +9,4 @@ import { economicCodes } from "../mockData/responseConstructor.js";
 export const isKey = (object, key) => {
     return key in object;
 };
-export const economicTables = Object.keys(economicCodes).map(key => `CODE_${key} INTEGER DEFAULT NULL, `).join("");
+export const economicTables = Object.keys(economicCodes).map(key => `CODE_${key} DOUBLE PRECISION DEFAULT NULL, `).join("");
